refactor: refatorando o sistema de acordo com o livro clean arch uncle bob
