+++ app/Domain/Entities/Manager/Manager.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Entities;
+
+class Manager
+{
+
+}
