+++ app/Domain/Entities/Professor/Professor.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Entities;
+
+final class Professor
+{
+
+}
