+++ app/Domain/Entities/Student/Student.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Entities;
+
+final class Student
+{
+
+}
