+++ app/Domain/Entities/User/User.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Entities;
+
+class User
+{
+
+}
