+++ app/Domain/Entities/Video/Video.php
@@ -1,0 +1,6 @@
+<?php
+
+class Video
+{
+
+}
